requirements.txt
flask
requests
elevenlabs==0.2.27
SpeechRecognition
pyaudio
python-dotenv